rss_gen::data
Struct RssItem

#[non_exhaustive]
pub struct RssItem {
    pub guid: String,
    pub description: String,
    pub link: String,
    pub pub_date: String,
    pub title: String,
    pub author: String,
    pub category: Option<String>,
    pub comments: Option<String>,
    pub enclosure: Option<String>,
    pub source: Option<String>,
}

Represents an item in the RSS feed.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

guid: String

The GUID of the RSS item (unique identifier).

description: String

The description of the RSS item.

link: String

The link to the RSS item.

pub_date: String

The publication date of the RSS item.

title: String

The title of the RSS item.

author: String

The author of the RSS item.

category: Option<String>

The category of the RSS item (optional).

comments: Option<String>

The comments URL related to the RSS item (optional).

enclosure: Option<String>

The enclosure (typically for media like podcasts) (optional).

source: Option<String>

The source of the RSS item (optional).

Implementations

impl RssItem

pub fn new() -> Self

Creates a new RssItem with default values.

pub fn set<T: Into<String>>(self, key: &str, value: T) -> Self

Sets the value of a field and returns the RssItem instance for method chaining.

Arguments

key - The field to set.

value - The value to assign to the field.

pub fn validate(&self) -> Result<()>

Validates the RssItem to ensure all required fields are set and valid.

pub fn guid<T: Into<String>>(self, value: T) -> Self

Sets the GUID.

pub fn description<T: Into<String>>(self, value: T) -> Self

Sets the description.

pub fn link<T: Into<String>>(self, value: T) -> Self

Sets the link.

pub fn pub_date<T: Into<String>>(self, value: T) -> Self

Sets the publication date.

pub fn title<T: Into<String>>(self, value: T) -> Self

Sets the title.

pub fn author<T: Into<String>>(self, value: T) -> Self

Sets the author.

pub fn pub_date_parsed(&self) -> Result<DateTime>

Parses the pub_date string into a DateTime object.

Returns

Ok(DateTime) if the date is valid and successfully parsed.

Err(RssError) if the date is invalid or cannot be parsed.