rss_gen::generator
Function write_element

pub fn write_element<W: Write>(
    writer: &mut Writer<W>,
    name: &str,
    content: &str,
) -> Result<()>

Writes an XML element with the given name and content.

Arguments

writer - A mutable reference to the XML writer.

name - The name of the XML element.

content - The content of the XML element.

Returns

A Result indicating success or failure of the write operation.

Errors

This function returns an Err if there is an issue with writing XML content.