rss_gen::validator
Function validate_rss_feed

pub fn validate_rss_feed(rss_data: &RssData) -> Result<()>

Validates the provided RssData and returns a Result indicating success or failure.

Arguments

rss_data - A reference to the RssData to be validated.

Returns

Ok(()) if the validation passes.

Err(RssError::ValidationErrors) containing a list of validation errors if any are found.

Errors

This function returns an Err(RssError::ValidationErrors) if any validation checks fail.